Successfully rendering a rotating wireframe cube using Besenham's Line Algorithm!

+++ RendererV1/src/renderer/Point2D.java
@@ -1,10 +1,10 @@
 package renderer;
 
 public class Point2D {
-    float x;
-    float y;
+    double x;
+    double y;
 
-    public Point2D(float x, float y) {
+    public Point2D(double x, double y) {
         this.x = x;
         this.y = y;
     }
